+++ backend/fetchPositionsFromSRAPI.js
@@ -1,13 +1,13 @@
 const { fetch } = require('wix-fetch');
 const { items: wixData } = require('@wix/data');
 const { COLLECTIONS } = require('./collectionConsts');
-const { getCompanyId } = require('./data');
+const { getCompanyId } = require('./secretsData');
 async function makeSmartRecruitersRequest(path,token) {
    const baseUrl = 'https://api.smartrecruiters.com'; // PROD
 //  const baseUrl = 'https://bayank2.wixstudio.com/my-site-3//_functions'; // TEST
   const fullUrl = `${baseUrl}${path}`;
   
-    //console.log(`Making request to: ${fullUrl}`);
+    //console.log(`Making request to: ${fullUrl}`); 
   try {
     const response = await fetch(fullUrl, {
       method: 'GET',
